Introduce a basic task runner system designed as a CLI
After a brief affair with Haku, I decided that I wanted to rewwrite my
build tasks in Rust. Surprisingly I couldn't find a single project that
simultaneously fit my needs and worked properly in a Cargo workspace.

SO, I wrote a little clap-rs based CLI project — cargo-task.

It's nothing fancy, but I'm feeling a lot happier moving to a workflow
where I'm able to reuse code effectively. Haku was originally appealing
because it didn't fall into the same traps as some of the other task
runners , but ultimately, the elements of reuse were not powerful enough
to satisfy me.